requirements.txt
APScheduler==3.2.0
bottle==0.12.10
gunicorn==19.6.0
httplib2==0.9.2
oauthlib==2.0.0
python-instagram==1.3.2
pytz==2016.7
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
simplejson==3.8.2
six==1.10.0
tweepy==3.5.0
tzlocal==1.3
